requirements.txt
arrow==0.7.0
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.4.1
binaryornot==0.4.0
certifi==2016.8.8
chardet==2.3.0
click==6.6
colorama==0.3.7
common==0.1.2
configparser==3.5.0
cookiecutter==1.4.0
cookiejar==0.0.1
coreapi==2.0.4
decorator==4.0.10
Django==1.9.5
django-appconf==1.0.1
django-autofixture==0.12.1
django-categories==1.4.1
django-chartit==0.2.7
django-ckeditor==5.0.3
django-cors-headers==1.1.0
django-cors-middleware==1.3.1
django-countries==3.4.1
django-enumfield==1.3b2
django-extensions==1.6.7
django-filter==0.14.0
django-google-charts==0.1.1
django-imagekit==3.3
django-markdown==0.8.4
django-material==0.8.0
django-model-utils==2.5
django-mptt==0.8.5
django-mptt-admin==0.4.0
django-related-admin==0.3.1
django-rest-swagger==2.0.5
djangorestframework==3.4.0
DjangoSEO==1.0
easy-thumbnails==2.3
entrypoints==0.2.2
enum34==1.1.6
fake-factory==0.6.0
functools32==3.2.3.post2
future==0.15.2
graphviz==0.4.10
httpie==0.9.6
ipaddress==1.0.16
ipykernel==4.4.1
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
itypes==1.1.0
Jinja2==2.8
jinja2-time==0.1.0
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.3.0
jupyter-console==5.0.0
jupyter-core==4.1.1
Markdown==2.6.6
MarkupSafe==0.23
mistune==0.7.3
mysqlclient==1.3.7
nbconvert==4.2.0
nbformat==4.1.0
nltk==3.2.1
notebook==4.2.2
numpy==1.11.0
openapi-codec==1.0.0
pager==3.3
panda==0.3.1
pandas==0.18.1
pathlib2==2.1.0
pickleshare==0.7.4
pilkit==1.1.13
Pillow==3.2.0
poyo==0.1.0
prompt-toolkit==1.0.7
psycopg2==2.6.2
pydot==1.1.0
Pygments==2.1.3
pyparsing==2.1.5
python-dateutil==2.5.2
pytz==2016.4
pyzmq==15.4.0
qtconsole==4.2.1
requests==2.9.1
simplegeneric==0.8.1
simplejson==3.8.2
singledispatch==3.4.0.3
six==1.10.0
tornado==4.4.1
traitlets==4.2.2
ujson==1.35
unicode-slugify==0.1.1
uritemplate==3.0.0
urlopen==1.0.0
wcwidth==0.1.7
whichcraft==0.1.1
widgetsnbextension==1.2.6
win-unicode-console==0.5
